// admin/src/api/generated/wheretolive.ts
export type ListAuditEventsParams = {
  /**
   * @maxLength 128
   */
  action?: string;
  /**
   * @maxLength 128
   */
  resource?: string;
  /**
   * @minimum 1
   * @maximum 100
   */
  limit?: number;
  /**
   * @minimum 0
   */
  offset?: number;
};